Made Enemy a Super class + changes combat system
Made enemy a super class and added 5 subclasses in order to give every single one of them a different weapon
Changed combat system, weapons assigned in the Players or Enemies class are now used and will do a random amount of damage depending on the minDamage value and the maxDamage value.
Added colors to the print lines.